requirements.txt
google.cloud
jsonschema
requests
enum